Move files into separate files
- breakout type and configuration definitions into Types.jl
- breakout basic actions to Actions.jl
- Add two missing arguments to function (before: from scope)
