+++ src/PlutoSliderServer.jl
@@ -5,6 +5,8 @@ using FromFile
 @from "./MoreAnalysis.jl" import MoreAnalysis
 @from "./FileHelpers.jl" import FileHelpers: find_notebook_files_recursive, list_files_recursive
 @from "./Export.jl" using Export
+@from "./Actions.jl" import Actions: add_to_session!, generate_static_export, myhash, showall
+@from "./Types.jl" using Types
 
 import Pluto
 import Pluto: ServerSession, Firebasey, Token, withtoken, pluto_file_extensions, without_pluto_file_extension
@@ -19,67 +21,6 @@ using Logging: global_logger
 using GitHubActions: GitHubActionsLogger
 function __init__()
     get(ENV, "GITHUB_ACTIONS", "false") == "true" && global_logger(GitHubActionsLogger())
-end
-showall(xs) = Text(join(string.(xs),"\n"))
-
-myhash = base64encode ∘ sha256
-
-###
-# SESSION DEFINITION
-
-abstract type NotebookSession end
-
-Base.@kwdef struct RunningNotebookSession <: NotebookSession
-    path::String
-    hash::String
-    notebook::Pluto.Notebook
-    original_state
-    token::Token=Token()
-    bond_connections::Dict{Symbol,Vector{Symbol}}
-end
-
-Base.@kwdef struct QueuedNotebookSession <: NotebookSession
-    path::String
-    hash::String
-end
-
-Base.@kwdef struct FinishedNotebookSession <: NotebookSession
-    path::String
-    hash::String
-    original_state
-end
-
-
-
-###
-# CONFIGURATION
-
-@option struct SliderServerSettings
-    exclude::Vector=String[]
-    port::Integer=2345
-    host="127.0.0.1"
-    simulated_lag::Real=0
-    serve_static_export_folder::Bool=true
-end
-
-@option struct ExportSettings
-    output_dir::Union{Nothing,String}=nothing
-    exclude::Vector=String[]
-    ignore_cache::Vector=String[]
-    pluto_cdn_root::Union{Nothing,String}=nothing
-    baked_state::Bool=true
-    offer_binder::Bool=true
-    disable_ui::Bool=true
-    cache_dir::Union{Nothing,String}=nothing
-    slider_server_url::Union{Nothing,String}=nothing
-    binder_url::Union{Nothing,String}=nothing
-    create_index::Bool=true
-end
-
-@option struct PlutoDeploySettings
-    SliderServer::SliderServerSettings=SliderServerSettings()
-    Export::ExportSettings=ExportSettings()
-    Pluto::Pluto.Configuration.Options=Pluto.Configuration.Options()
 end
 
 function get_configuration(toml_path::Union{Nothing,String}=nothing; kwargs...)
@@ -260,137 +201,16 @@ function run_directory(
         
         @info "[$(i)/$(length(to_run))] Opening $(path)"
 
-        function add_to_session!(notebook_sessions, server_session, path, settings)
-        jl_contents = read(joinpath(start_dir, path), String)
-        hash = myhash(jl_contents)
-        # The 5 lines below are needed if the code is not invoked within PlutoSliderServer.run_directory
-        i = findfirst(s -> s.hash == hash, notebook_sessions)
-        if isnothing(i)
-            push!(notebook_sessions, QueuedNotebookSession(;path, hash=hash))
-            i = length(notebook_sessions)
-        end
-        keep_running = run_server && path ∉ settings.SliderServer.exclude
-        skip_cache = keep_running || path ∈ settings.Export.ignore_cache
-
-        local notebook, original_state
-        
-        cached_state = skip_cache ? nothing : try_fromcache(settings.Export.cache_dir, hash)
-        if cached_state !== nothing
-            @info "Loaded from cache, skipping notebook run" hash
-            original_state = cached_state
-        else
-            try
-                # open and run the notebook (TODO: tell pluto not to write to the notebook file)
-                notebook = Pluto.SessionActions.open(server_session, joinpath(start_dir, path); run_async=false)
-                # get the state object
-                original_state = Pluto.notebook_to_js(notebook)
-                # shut down the notebook
-                if !keep_running
-                    @info "Shutting down notebook process"
-                    Pluto.SessionActions.shutdown(server_session, notebook)
-                end
-                if keep_running
-                    bond_connections = MoreAnalysis.bound_variable_connections_graph(notebook)
-                    @info "Bond connections" showall(collect(bond_connections))
-
-                    # By setting notebook_sessions[i] to a running session, (modifying the array), the HTTP router will now start serving requests for this notebook.
-                    notebook_sessions[i] = RunningNotebookSession(;
-                        path,
-                        hash,
-                        notebook,
-                        original_state,
-                        bond_connections,
-                    )
-                else
-                    notebook_sessions[i] = FinishedNotebookSession(;
-                        path,
-                        hash,
-                        original_state,
-                    )
-                end
-                try_tocache(settings.Export.cache_dir, hash, original_state)
-            catch e
-                (e isa InterruptException) || rethrow(e)
-                @error "Failed to run notebook!" path exception=(e,catch_backtrace())
-                # continue
-            end
-        end
-        notebook_sessions[i], jl_contents, original_state
-    end
-
     local session, jl_contents, original_state
     # That's because you can't continue in a loop
     try
-        session, jl_contents, original_state = add_to_session!(notebook_sessions, server_session, path, settings)
+        session, jl_contents, original_state = add_to_session!(notebook_sessions, server_session, path, settings, run_server, start_dir)
     catch e
         rethrow(e)
         continue
     end
 
     if static_export
-        function generate_static_export(path, settings, original_state=nothing, output_dir=".", jl_contents=nothing)
-            export_jl_path = let
-                relative_to_notebooks_dir = path
-                joinpath(output_dir, relative_to_notebooks_dir)
-            end
-            export_html_path = let
-                relative_to_notebooks_dir = without_pluto_file_extension(path) * ".html"
-                joinpath(output_dir, relative_to_notebooks_dir)
-            end
-            export_statefile_path = let
-                relative_to_notebooks_dir = without_pluto_file_extension(path) * ".plutostate"
-                joinpath(output_dir, relative_to_notebooks_dir)
-            end
-
-
-            mkpath(dirname(export_jl_path))
-            mkpath(dirname(export_html_path))
-            mkpath(dirname(export_statefile_path))
-
-
-            notebookfile_js = if (settings.Export.offer_binder || settings.Export.slider_server_url !== nothing)
-                repr(basename(export_jl_path))
-            else
-                "undefined"
-            end
-            slider_server_url_js = if settings.Export.slider_server_url !== nothing
-                repr(settings.Export.slider_server_url)
-            else
-                "undefined"
-            end
-            binder_url_js = if settings.Export.offer_binder
-                repr(something(settings.Export.binder_url, "https://mybinder.org/v2/gh/fonsp/pluto-on-binder/v$(string(pluto_version))"))
-            else
-                "undefined"
-            end
-            statefile_js = if !settings.Export.baked_state
-                open(export_statefile_path, "w") do io
-                    Pluto.pack(io, original_state)
-                end
-                repr(basename(export_statefile_path))
-            else
-                statefile64 = base64encode() do io
-                    Pluto.pack(io, original_state)
-                end
-
-                "\"data:;base64,$(statefile64)\""
-            end
-
-            html_contents = generate_html(;
-                pluto_cdn_root=settings.Export.pluto_cdn_root,
-                version=pluto_version,
-                notebookfile_js, statefile_js,
-                slider_server_url_js, binder_url_js,
-                disable_ui=settings.Export.disable_ui
-            )
-            write(export_html_path, html_contents)
-
-            if (settings.Export.offer_binder || settings.Export.slider_server_url !== nothing)
-                write(export_jl_path, jl_contents)
-            end
-
-            @info "Written to $(export_html_path)"
-            end
             generate_static_export(path, settings, original_state, output_dir, jl_contents)
         end
